requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyMuPDF
Pillow
pytesseract
langchain
langchain-mistralai
sentence-transformers
faiss-cpu
elasticsearch
celery
redis
python-multipart
itsdangerous
email-validator
typing_extensions
aioboto3
boto3